linux client test done

+++ client.h
@@ -71,7 +71,7 @@ long long getSystemTime();
  *  9. file list ()
  * */
 typedef struct {
-    unsigned long timeStamp;// length: 13
+    long long timeStamp;// length: 13
     unsigned short ptype; // [4 - 9]
     unsigned short fmode; // 1: On, 0: Off
     unsigned short lmode; // 1: On, 0: Off
@@ -83,7 +83,7 @@ typedef struct {
 } Request;
 
 typedef struct {
-    unsigned long timeStamp; // length :13
+    long long timeStamp; // length :13
     unsigned short ptype; // [4 - 9]
     unsigned short success; // 1/0 cmd execute success or not
     char* response; // response message
